+++ sherpa/cpp_api/websocket/online-websocket-server-impl.cc
@@ -18,7 +18,7 @@
 
 #include "sherpa/cpp_api/websocket/online-websocket-server-impl.h"
 
-#include "sherpa/csrc/file_utils.h"
+#include "sherpa/csrc/file-utils.h"
 #include "sherpa/csrc/log.h"
 
 namespace sherpa {
