+++ sherpa/csrc/fbank-features.h
