+++ sherpa/csrc/symbol-table.h
